# HARMONY-1932: Fix STAC item request hanging.