--- a/db/migrations/20241021192628_add_jobid_id_index_to_job_links_table.js
+++ b/db/migrations/20241021192628_add_jobid_id_index_to_job_links_table.js
@@ -1,0 +1,21 @@
+/**
+ * @param { import("knex").Knex } knex
+ * @returns { Promise<void> }
+ */
+exports.up = function(knex) {
+  return knex.schema
+    .table('job_links', function (table) {
+      table.index(['jobID', 'id']);
+    });
+};
+
+/**
+ * @param { import("knex").Knex } knex
+ * @returns { Promise<void> }
+ */
+exports.down = function(knex) {
+  return knex.schema
+    .table('job_links', function (table) {
+      table.dropIndex(['jobID', 'id']);
+    });
+};