## Jira Issue ID
HARMONY-1932

## Description
This PR fixes STAC item request hanging by adding a composite index on job_links table so that the STAC item request db query will use the composite index instead of the expensive table scan. After adding the index, the db query becomes tens of thousands times faster. See below for the query plan before and after the introduction of index:

### Before adding index:
```
Limit  (cost=0.56..3561.64 rows=1 width=337) (actual time=428998.478..428998.480 rows=1 loops=1)
  ->  LockRows  (cost=0.56..11028643.47 rows=3097 width=337) (actual time=428998.476..428998.477 rows=1 loops=1)
        ->  Index Scan using job_links_pkey on job_links  (cost=0.56..11028612.50 rows=3097 width=337) (actual time=428997.525..428997.526 rows=1 loops=1)
              Filter: ((bbox IS NOT NULL) AND ("temporalStart" IS NOT NULL) AND ("temporalEnd" IS NOT NULL) AND ("jobID" = 'cb2ce1fb-e0da-4111-b1a6-348da233fd2a'::uuid) AND ((rel)::text = 'data'::text))
              Rows Removed by Filter: 42441272
Planning Time: 15.293 ms
Execution Time: 428998.526 ms
```
### After adding index:
```
Limit  (cost=0.56..4.74 rows=1 width=337) (actual time=9.922..9.924 rows=1 loops=1)
  ->  LockRows  (cost=0.56..12919.43 rows=3097 width=337) (actual time=9.921..9.922 rows=1 loops=1)
        ->  Index Scan using job_links_jobid_id_index on job_links  (cost=0.56..12888.46 rows=3097 width=337) (actual time=6.305..6.306 rows=1 loops=1)
              Index Cond: ("jobID" = 'cb2ce1fb-e0da-4111-b1a6-348da233fd2a'::uuid)
              Filter: ((bbox IS NOT NULL) AND ("temporalStart" IS NOT NULL) AND ("temporalEnd" IS NOT NULL) AND ((rel)::text = 'data'::text))
              Rows Removed by Filter: 1
Planning Time: 12.231 ms
Execution Time: 9.986 ms
```
## Local Test Steps
I restored PROD database which has over 42 million rows in job_links table in Sandbox. Configured my Harmony service in Sandbox to use the new db cluster. I am able to reproduce the problem in Sandbox and verify after adding the index, the problem is resolved. See: https://internal-harmony-yliu10-frontend-1173703343.us-west-2.elb.amazonaws.com/stac/ae866fa5-8646-4eff-8d32-1c5440a8ad83/1

## PR Acceptance Checklist
* [x] Acceptance criteria met
* [ ] Tests added/updated (if needed) and passing
* [ ] Documentation updated (if needed)